requirements.txt
python-dotenv
requests
pydantic
numpy
